requirements.txt
numpy
pandas
pytorch
torchvision
matplotlib
scipy
skimage